libct/state: remove double call to runType
In case refreshState is called for container with the current state of
Running but which has exited, refreshState calls runType (which returns
Stopped), and when calls (*runningState).transition(&stoppedState),
which calls runType() again.

Since refreshState is the only caller of
(*runningState).transition(&stoppedState), the second call to runType is
redundant. Remove it.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>