Pass a pointer to a struct where possible
I have noticed that newProcess function in utils_linux.go
receives spec.Process argument (rather than a pointer to it).
It means that the whole structure is being copied, and this structure
is not a small one -- its size is a few hundred bytes.

Surely, we can just pass a pointer to it.

Next, I found gocritic that has this type of check (named hugeParam),
and found a few more cases like the one above. Alas, some of the issues
found could not be fixed without some refactoring, so this commit only
fixes one more such case -- methods of libcontainer.Process.

This fixes the following gocritic hugeParam warnings:

./utils_linux.go:79:17: hugeParam: p is heavy (232 bytes); consider passing it by pointer
./libcontainer/process.go:96:7: hugeParam: p is heavy (312 bytes); consider passing it by pointer
./libcontainer/process.go:104:7: hugeParam: p is heavy (312 bytes); consider passing it by pointer
./libcontainer/process.go:114:7: hugeParam: p is heavy (312 bytes); consider passing it by pointer

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>